requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
schedule>=1.2.0
<<<<<<< Updated upstream
=======
pymysql>=1.1.0
sshtunnel>=0.4.0
cryptography>=41.0.0
>>>>>>> Stashed changes
